Backend/requirements.txt
alembic==1.16.5
aniso8601==10.0.1
bcrypt==4.3.0
blinker==1.9.0
click==8.3.0
Faker==19.13.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==6.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
PyMySQL==1.1.2
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.1.2